fix(backend): Hibernate 7 type strictness + SB4 bean wiring fixes
Three issues found during live verification:

1. SMALLINT vs INTEGER: Hibernate 7 validates JDBC type codes strictly
   (int2 ≠ int4). Changed sectionNumber fields from Integer/Short in
   AnnexIvSection and Stage2Submission to Short, and matching repo
   method signatures.

2. RestClient.Builder not auto-configured in Spring Boot 4: removed
   constructor injection, now uses RestClient.builder() static factory
   directly in VoyageEmbeddingClient.

3. ObjectMapper not auto-configured in ingest profile context:
   replaced constructor injection with a private static final
   OBJECT_MAPPER instance in both ingestion services.

All three verified: cold start passes ddl-auto:validate, ./mvnw test
passes (1/1), ingest profile fires and halts correctly on Voyage 429
(rate limit — requires billing setup, not a code issue).

+++ backend/src/main/java/com/regu/domain/AnnexIvSection.java
@@ -28,9 +28,11 @@ public class AnnexIvSection {
     /**
      * Annex IV section number (1–9). UNIQUE in the database.
      * Referenced directly by {@code stage2_submissions.section_number}.
+     * Mapped as {@code Short} to match the {@code SMALLINT} column type —
+     * Hibernate 7 validates JDBC type codes strictly (int2 ≠ int4).
      */
     @Column(name = "section_number", nullable = false)
-    private Integer sectionNumber;
+    private Short sectionNumber;
 
     @Column(name = "section_key", nullable = false, length = 40)
     private String sectionKey;
@@ -70,7 +72,7 @@ public class AnnexIvSection {
 
     protected AnnexIvSection() { }
 
-    public AnnexIvSection(Integer sectionNumber, String sectionKey,
+    public AnnexIvSection(Short sectionNumber, String sectionKey,
                           String displayTitle, String instructions) {
         this.sectionNumber = sectionNumber;
         this.sectionKey = sectionKey;
@@ -85,8 +87,8 @@ public class AnnexIvSection {
     public UUID getId() { return id; }
     public void setId(UUID id) { this.id = id; }
 
-    public Integer getSectionNumber() { return sectionNumber; }
-    public void setSectionNumber(Integer sectionNumber) { this.sectionNumber = sectionNumber; }
+    public Short getSectionNumber() { return sectionNumber; }
+    public void setSectionNumber(Short sectionNumber) { this.sectionNumber = sectionNumber; }
 
     public String getSectionKey() { return sectionKey; }
     public void setSectionKey(String sectionKey) { this.sectionKey = sectionKey; }
